requirements.txt
# Core libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=1.5.0
pillow>=9.5.0

# Deep learning and model related
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0

# Computer vision
opencv-python>=4.7.0
pycocotools>=2.0.6
scikit-image>=0.19.0

# Data handling and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tqdm>=4.65.0

# Labelbox integration
labelbox>=3.0.0

# Logging and monitoring
wandb>=0.15.0

# Utilities
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0
setuptools>=65.0.0
wheel>=0.38.0

# Development tools (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 